work on generating ace report

+++ tools/mkepub.sh
@@ -24,14 +24,16 @@ mv Book.zip ContraCostaCountyFieldGuide.epub
 sh ../tools/epubcheck.sh ContraCostaCountyFieldGuide.epub
 
 if hash ace 2>/dev/null; then
-  ace -f -s -o AceReport ContraCostaCountyFieldGuide.epub
-  rm -rf ${CWD}/AceReport/data
-  [ ! -d ${CWD}/AceReport ] && mkdir ${CWD}/AceReport
-  mv AceReport/data ${CWD}/AceReport/
-  mv AceReport/report.html ${CWD}/AceReport/
-  mv AceReport/report.json ${CWD}/AceReport/
-  echo "Accessibility report written to AceReport directory"
-  echo `pwd`
+  if [ ! -f ${CWD}/AceReport/noace.tmp ]; then
+    ace -f -s -o AceReport ContraCostaCountyFieldGuide.epub
+    rm -rf ${CWD}/AceReport/data
+    [ ! -d ${CWD}/AceReport ] && mkdir ${CWD}/AceReport
+    mv AceReport/data ${CWD}/AceReport/
+    mv AceReport/report.html ${CWD}/AceReport/
+    mv AceReport/report.json ${CWD}/AceReport/
+    echo "Accessibility report written to AceReport directory"
+    echo `pwd`
+  fi
 fi
 
 
@@ -41,7 +43,9 @@ popd
 
 if hash ace 2>/dev/null; then
   if [ -f AceReport/.gitignore ]; then
-    git commit -m "update AceReport" AceReport/report.*
+    if [ ! -f AceReport/noace.tmp ]; then
+      git commit -m "update AceReport" AceReport/report.*
+    fi
   fi
 fi
